purge: allow custom purge command.
Allow defining a purge command (parameter purge_parameters::purge_command)
instead of performing 'unlink' action as built-in.
Purge command can take custom parameters like {path}, {fsname}, {fid}.
Added non-reg tests for this feature.

+++ src/policy_modules/purge/resmon_purge.c
@@ -76,16 +76,48 @@ static int PurgeEntry_ByFid( const entry_id_t * p_entry_id,
 }
 #else
 /** purge by path for other purpose */
-static int PurgeEntry_ByPath( const char *entry_path )
+static int PurgeEntry(const entry_id_t *id, const char *entry_path)
 {
-    DisplayLog( LVL_FULL, PURGE_TAG, "Unlink(%s)", entry_path );
-    if ( !dry_run )
+    if (EMPTY_STRING(resmon_config.purge_command))
     {
-        if ( unlink( entry_path ) != 0 )
+        /* no custom purge command => unlink */
+        DisplayLog(LVL_DEBUG, PURGE_TAG, "%sunlink(%s)", dry_run ? "(dry-run) " : "",
+                   entry_path);
+        if (!dry_run)
+        {
+            if (unlink( entry_path) != 0)
+                return errno;
+        }
+        return 0;
+    }
+    else
+    {
+        /* execute custom action */
+        char strfid[128];
+        sprintf(strfid, DFID, PFID(id));
+
+        const char *vars[] = {
+            "path", entry_path,
+            "fsname", get_fsname(),
+            "fid", strfid,
+            NULL, NULL
+        };
+
+        char *cmd = replace_cmd_parameters(resmon_config.purge_command, vars);
+        if (cmd)
+        {
+            int rc = 0;
+            /* call custom purge command instead of unlink() */
+            DisplayLog(LVL_DEBUG, PURGE_TAG, "%scmd(%s)", dry_run ? "(dry-run) " : "", cmd);
+            if (!dry_run)
+                rc =  execute_shell_command(cmd, 0);
+            free(cmd);
+            return rc;
+        }
+        else
             return errno;
         /* @TODO handle other hardlinks to the same entry */
     }
-    return 0;
 }
 #endif
 
@@ -1385,7 +1417,7 @@ static void ManageEntry( lmgr_t * lmgr, purge_item_t * p_item )
     }
 #else
     /* FIXME should remove all paths to the object */
-    rc = PurgeEntry_ByPath( ATTR( &new_attr_set, fullpath ) );
+    rc = PurgeEntry(&p_item->entry_id, ATTR(&new_attr_set, fullpath));
 #endif
     if ( rc )
     {
